authenticate: fix expiring user info endpoint
